Revert "Enchance UI and Auth Bloc"
