--- lib/screens/home.dart
@@ -1,187 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:login/screens/job_detail.dart';
-import 'package:login/screens/about_job.dart';
-import 'package:login/screens/notif_page.dart';
-import 'package:login/screens/payment_page.dart';
-import 'package:login/screens/post_page.dart';
-import 'package:login/screens/profile_page.dart';
-// import 'balance_card.dart';
-import '../widget/home/job_card.dart';
-import '../widget/home/search_and_filter.dart';
-import '../widget/home/section_title.dart';
-import '../widget/home/section_title_with_dropdown.dart';
-
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Row(
-          children: [
-            SizedBox(width: 16),
-            Text(
-              'Home',
-              style: TextStyle(
-                fontSize: 18,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-            ),
-          ],
-        ),
-        backgroundColor: const Color.fromARGB(255, 3, 169, 244),
-        automaticallyImplyLeading: false,
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.notifications, color: Colors.white),
-            onPressed: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                    builder: (context) => const NotificationPage()),
-              );
-            },
-          ),
-        ],
-      ),
-      backgroundColor: const Color.fromARGB(255, 255, 255, 255),
-      body: CustomScrollView(
-        slivers: [
-          SliverToBoxAdapter(
-            child: Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  const SearchAndFilter(),
-                  const SizedBox(height: 20),
-                  // const HomeBalanceCard(),
-                  const SizedBox(height: 20),
-                  const HomeSectionTitle(title: 'Most Recent Job'),
-                  const SizedBox(height: 10),
-                  SizedBox(
-                    height: 150,
-                    child: ListView.builder(
-                      scrollDirection: Axis.horizontal,
-                      itemCount: 10,
-                      itemBuilder: (context, index) {
-                        return Padding(
-                          padding: const EdgeInsets.only(right: 10),
-                          child: HomeJobCard(
-                            jobTitle: 'Recent Job Title ${index + 1}',
-                            jobDescription:
-                                'Details of the most recent job ${index + 1}.',
-                            showAddButton: false,
-                            onTap: () {
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) => const JobDetailPage(
-                                          jobTitle: '',
-                                          jobDescription: '',
-                                        )),
-                              );
-                            },
-                          ),
-                        );
-                      },
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                  const SectionTitleWithDropdown(title: 'Recommended'),
-                  const SizedBox(height: 10),
-                ],
-              ),
-            ),
-          ),
-          SliverList(
-            delegate: SliverChildBuilderDelegate(
-              (context, index) {
-                return Padding(
-                  padding: const EdgeInsets.only(bottom: 16.0),
-                  child: HomeJobCard(
-                    jobTitle: 'Job Category Title',
-                    jobDescription: 'Job details by category.',
-                    showAddButton: true,
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                            builder: (context) => const JobDetailPage(
-                                  jobTitle: '',
-                                  jobDescription: '',
-                                )),
-                      );
-                    },
-                  ),
-                );
-              },
-              childCount: 10,
-            ),
-          ),
-        ],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        type: BottomNavigationBarType.fixed,
-        backgroundColor: const Color.fromARGB(255, 255, 255, 255),
-        currentIndex: 0,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.home_outlined),
-            label: 'Home',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.info_outline),
-            label: 'About Job',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.post_add),
-            label: 'Post',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.payment_outlined),
-            label: 'Payment',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.person_2_outlined),
-            label: 'Profile',
-          ),
-        ],
-        selectedItemColor: Colors.black,
-        unselectedItemColor: Colors.lightBlue,
-        onTap: (index) {
-          switch (index) {
-            case 0:
-              break;
-            case 1:
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => const JobPage()),
-              );
-              break;
-            case 2:
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => const PostPage()),
-              );
-              break;
-            case 3:
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => const PaymentPage()),
-              );
-              break;
-            case 4:
-              Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => const ProfilePage()),
-              );
-              break;
-          }
-        },
-      ),
-    );
-  }
-}
