--- lib/screens/post_page.dart
@@ -1,120 +1,0 @@
-import 'dart:io';
-import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import '../widget/post_page/post_page_ui.dart';
-
-class PostPage extends StatefulWidget {
-  const PostPage({super.key});
-
-  @override
-  PostPageState createState() => PostPageState();
-}
-
-class PostPageState extends State<PostPage> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _locationController = TextEditingController();
-  final TextEditingController _professionController = TextEditingController();
-  final TextEditingController _rewardController = TextEditingController();
-  final TextEditingController _descriptionController = TextEditingController();
-  final TextEditingController _contactController = TextEditingController();
-  final TextEditingController _emailController = TextEditingController();
-
-  final List<XFile> _images = [];
-
-  Future<void> _pickImages() async {
-    final ImagePicker picker = ImagePicker();
-    final List<XFile>? images = await picker.pickMultiImage();
-    if (images != null) {
-      setState(() {
-        _images.addAll(images);
-      });
-    }
-  }
-
-  void _removeImage(int index) {
-    setState(() {
-      _images.removeAt(index);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          'Post a New Job',
-          style: TextStyle(
-            fontSize: 20,
-            fontWeight: FontWeight.bold,
-            color: Colors.white,
-          ),
-        ),
-        backgroundColor: Colors.lightBlueAccent,
-        elevation: 0,
-        centerTitle: true,
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back, color: Colors.white),
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-      ),
-      backgroundColor: const Color.fromARGB(255, 255, 255, 255),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(5.0),
-            child: Form(
-              key: _formKey,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  buildTextField(_locationController, 'Location'),
-                  const SizedBox(height: 16),
-                  buildTextField(_professionController, 'Profession'),
-                  const SizedBox(height: 16),
-                  buildTextField(_rewardController, 'Reward'),
-                  const SizedBox(height: 16),
-                  buildTextField(
-                    _descriptionController,
-                    'Description',
-                    maxLines: 5,
-                  ),
-                  const SizedBox(height: 16),
-                  buildTextField(_contactController, 'Contact Number'),
-                  const SizedBox(height: 16),
-                  buildTextField(_emailController, 'Email Address'),
-                  const SizedBox(height: 20),
-                  buildAttachmentSection(_images, _pickImages, _removeImage),
-                  const SizedBox(height: 20),
-                  Center(
-                    child: ElevatedButton(
-                      onPressed: () {
-                        if (_formKey.currentState?.validate() ?? false) {
-                          // Handle form submission
-                        }
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.lightBlueAccent,
-                        padding: const EdgeInsets.symmetric(
-                            vertical: 15, horizontal: 30),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                      ),
-                      child: const Text(
-                        'Post Job',
-                        style: TextStyle(fontSize: 18, color: Colors.white),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
