+++ lib/views/job_detail.dart
