+++ lib/views/notif_page.dart
@@ -89,8 +89,8 @@ class NotificationItem {
 
 List<NotificationItem> notifications = [
   NotificationItem(
-    title: 'Someone accepted your job offer',
-    subtitle: 'Check out your job offer.',
+    title: 'New Job Available',
+    subtitle: 'Check out the new job opportunity',
     message:
         'A new job in your preferred category has been posted. Check it out now!',
   ),
