--- lib/widget/home/job_card.dart
@@ -1,71 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomeJobCard extends StatelessWidget {
-  final String jobTitle;
-  final String jobDescription;
-  final VoidCallback onTap;
-  final bool showAddButton;
-
-  const HomeJobCard({
-    super.key,
-    required this.jobTitle,
-    required this.jobDescription,
-    required this.onTap,
-    this.showAddButton = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Card(
-        color: const Color.fromARGB(255, 255, 255, 255),
-        child: Stack(
-          children: [
-            SizedBox(
-              height: 100,
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      jobTitle,
-                      style: const TextStyle(
-                          fontSize: 16, fontWeight: FontWeight.bold),
-                    ),
-                    const SizedBox(height: 8),
-                    Text(
-                      jobDescription,
-                      style: const TextStyle(fontSize: 14),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            if (showAddButton)
-              Positioned(
-                right: 15,
-                top: 35,
-                child: Container(
-                  height: 30,
-                  width: 30,
-                  decoration: const BoxDecoration(
-                    color: Colors.lightBlue,
-                    shape: BoxShape.circle,
-                  ),
-                  child: const Center(
-                    child: Icon(
-                      Icons.add,
-                      color: Colors.white,
-                      size: 20,
-                    ),
-                  ),
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
