--- lib/widget/home/search_and_filter.dart
@@ -1,41 +1,0 @@
-// lib/views/home/search_and_filter.dart
-import 'package:flutter/material.dart';
-
-class SearchAndFilter extends StatelessWidget {
-  const SearchAndFilter({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(top: 0.0),
-      child: Row(
-        children: [
-          Expanded(
-            child: TextField(
-              style: const TextStyle(fontSize: 14),
-              decoration: InputDecoration(
-                hintText: 'Search',
-                hintStyle: const TextStyle(fontSize: 14),
-                prefixIcon: const Icon(Icons.search, size: 20),
-                contentPadding: const EdgeInsets.symmetric(vertical: 10.0),
-                filled: true,
-                fillColor: Colors.grey[200],
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(10),
-                  borderSide: BorderSide.none,
-                ),
-              ),
-            ),
-          ),
-          const SizedBox(width: 0),
-          IconButton(
-            icon: const Icon(Icons.filter_list, size: 20),
-            onPressed: () {
-              Scaffold.of(context).openDrawer();
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
