--- lib/widget/home/section_title.dart
@@ -1,16 +1,0 @@
-// lib/views/home/section_title.dart
-import 'package:flutter/material.dart';
-
-class HomeSectionTitle extends StatelessWidget {
-  final String title;
-
-  const HomeSectionTitle({super.key, required this.title});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      title,
-      style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-    );
-  }
-}
