--- lib/widget/payment/balance_card.dart
@@ -1,70 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BalanceCard extends StatelessWidget {
-  const BalanceCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 100,
-      width: MediaQuery.of(context).size.width * 0.89, // Responsive width
-      decoration: BoxDecoration(
-        color: Colors.transparent,
-        borderRadius: BorderRadius.circular(10),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 4,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: Stack(
-        children: [
-          Opacity(
-            opacity: 0.3,
-            child: Container(
-              height: 100,
-              width: double.infinity,
-              decoration: const BoxDecoration(
-                color: Color.fromARGB(255, 255, 255, 255),
-                borderRadius: BorderRadius.only(
-                  topRight: Radius.circular(10),
-                  bottomRight: Radius.circular(10),
-                  topLeft: Radius.circular(10),
-                  bottomLeft: Radius.circular(10),
-                ),
-              ),
-            ),
-          ),
-          const Padding(
-            padding: EdgeInsets.all(16.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Flexible(
-                  child: Text(
-                    'Current Balance:',
-                    style: TextStyle(fontSize: 15, color: Colors.white),
-                    overflow: TextOverflow.ellipsis, // Prevents overflow
-                  ),
-                ),
-                SizedBox(height: 4), // Add space between the texts
-                Flexible(
-                  child: Text(
-                    '₱ 23,587',
-                    style: TextStyle(
-                        fontSize: 20,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                    overflow: TextOverflow.ellipsis, // Prevents overflow
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
