--- lib/widget/profile/section_with_see_all.dart
@@ -1,36 +1,0 @@
-import 'package:flutter/material.dart';
-
-class SectionWithSeeAll extends StatelessWidget {
-  final String title;
-  final VoidCallback onSeeAll;
-
-  const SectionWithSeeAll(
-      {super.key, required this.title, required this.onSeeAll});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          title,
-          style: const TextStyle(
-            fontSize: 22,
-            fontWeight: FontWeight.bold,
-            color: Colors.black87,
-          ),
-        ),
-        TextButton(
-          onPressed: onSeeAll,
-          child: const Text(
-            'See All',
-            style: TextStyle(
-              fontSize: 16,
-              color: Colors.blueAccent,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
