+++ pubspec.yaml
@@ -31,8 +31,6 @@ dependencies:
   flutter:
     sdk: flutter
   image_picker: ^0.8.5+3
-  flutter_bloc: ^8.0.0
-  equatable: ^2.0.0
 
 
   # The following adds the Cupertino Icons font to your application.
